last fixes to a few methods
